pyproject.toml
##
# This file describes the project's metadata and dependencies in the
# standard PEP 621 format.  It enables modern tooling such as uv to
# manage and install dependencies.  When using uv, you can add or
# remove dependencies by editing the tables here or by running
# commands like `uv add` and `uv remove`.  See the uv documentation
# for details on dependency management【832199573188594†L166-L182】.

[project]
name = "langchain_agentic_services"
version = "0.1.0"
description = "A simple LangChain-like agent for translation and question answering using a local vLLM server."
authors = [
  { name = "Nec", email = "nec.zhang.711@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

# Core runtime dependencies.  These are required for the agent to run.
dependencies = [
  "requests>=2.31.0",
  "scikit-learn>=1.7.1",
  "python-pptx>=1.0.2",
  "PyMuPDF>=1.26.3",
  "Pillow>=11.3.0",
  "tqdm>=4.67.1",
  "fastapi>=0.116.1",
  "uvicorn>=0.35.0",
  "python-multipart>=0.0.20",
  "python-dotenv>=1.1.1",
  "numpy>=2.3.2",
  "scipy>=1.16.1",
  "joblib>=1.5.1",
  "python-docx>=1.2.0",
  "pytesseract>=0.3.13",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "alembic>=1.12.0",
]

# Optional dependencies for additional functionality
[project.optional-dependencies]
extras = [
  "openpyxl>=3.1.0",
  "xlrd>=2.0.1"
]
# Development dependencies for testing, linting, and code quality
dev = [
  "mypy>=1.5.0",
  "black>=23.0.0",
  "flake8>=6.0.0",
  "pytest>=7.4.0",
  "bandit>=1.8.0",
  "safety>=3.6.0"
]

[build-system]
requires = [
  "setuptools>=61.0"
]
build-backend = "setuptools.build_meta"

# Fix package discovery issue
[tool.setuptools.packages.find]
include = ["simple_agent*"]
exclude = ["docker*", "tests*", "*.tmp*"]
